caseRef from Integer to Long

+++ src/test/java/uk/gov/ons/census/action/builders/FieldworkFollowupBuilderTest.java
@@ -52,7 +52,7 @@ public class FieldworkFollowupBuilderTest {
     assertThat(actualResult.getUndeliveredAsAddress()).isEqualTo(caze.isUndeliveredAsAddressed());
     assertThat(actualResult.getBlankQreReturned()).isFalse();
     assertThat(actualResult.getCaseId()).isEqualTo(caze.getCaseId().toString());
-    assertThat(actualResult.getCaseRef()).isEqualTo(Integer.toString(caze.getCaseRef()));
+    assertThat(actualResult.getCaseRef()).isEqualTo(Long.toString(caze.getCaseRef()));
     assertThat(actualResult.getHandDelivery()).isEqualTo(caze.isHandDelivery());
     assertThat(actualResult)
         .isEqualToIgnoringGivenFields(
